fix api key handling
 * move API key handling from environment to get_run
 * clean up code related to getting API key from env outside of
   get_run()
