+++ data_validation.py
@@ -1,10 +1,22 @@
-from prefect import task, flow, get_run_logger
+import os
 import time as ttime
+
+from prefect import task, flow, get_run_logger
 from tiled.client import from_uri
+from dotenv import load_dotenv
+
+
+def get_api_key_from_env():
+    with open("/srv/container.secret", "r") as secrets:
+        load_dotenv(stream=secrets)
+    api_key = os.environ["TILED_API_KEY"]
+    return api_key
 
 
 @task
 def get_run(uid, api_key=None):
+    if not api_key:
+        api_key = get_api_key_from_env()
     tiled_client = from_uri("https://tiled.nsls2.bnl.gov", api_key=api_key)
     run = tiled_client["srx/raw"][uid]
     return run
